refactor(finder): extract pure helpers from useFileSystem into utils/fileSystemHelpers
